Move osrm_car_duration.h to includes directory
