+++ src/sif/CMakeLists.txt
@@ -7,7 +7,6 @@ set(sources
   motorcyclecost.cc
   motorscootercost.cc
   nocost.cc
-  osrm_car_duration.h
   pedestriancost.cc
   transitcost.cc
   truckcost.cc
